add 3 laba

+++ ppo4/fproduct/models.py
@@ -5,7 +5,7 @@ class Finishs(models.Model):
     Name = models.CharField(max_length=255)
     Unit_id = models.ForeignKey(Unit, on_delete=models.CASCADE, db_column='Unit_id')
     Quantity = models.FloatField()
-    Amount = models.DecimalField(max_digits=19, decimal_places=2)
+    Amount = models.FloatField()
 
     class Meta:
         db_table = 'FProducts'
